// foot5connect-ui/src/app/services/models/available-player-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AvailablePlayerDto {
  availableDate?: string;
  city?: string;
  country?: string;
  endTime?: string;
  firstName?: string;
  lastName?: string;
  level?: 'DEBUTANT' | 'INTERMEDIAIRE' | 'CONFIRMER' | 'AVANCE';
  startTime?: string;
  totalGoals?: number;
  totalMatches?: number;
  userId?: number;
}